+++ electron/src/electron-utils/utilsEncryption.ts
@@ -35,4 +35,29 @@ export class UtilsEncryption {
       );
     }
   }
+  public async decryptDatabase(
+    pathDB: string,
+    password: string,
+  ): Promise<void> {
+    const msg = 'DecryptDatabase: ';
+    const retB: boolean = this.fileUtil.isPathExists(pathDB);
+    if (retB) {
+      try {
+        const mDB = await this.sqliteUtil.openOrCreateDatabase(
+          pathDB,
+          password,
+          false,
+        );
+        this.sqliteUtil.pragmaReKey(mDB, password, '');
+        this.sqliteUtil.closeDB(mDB);
+        return Promise.resolve();
+      } catch (err) {
+        return Promise.reject(new Error(`${msg} ${err.message} `));
+      }
+    } else {
+      return Promise.reject(
+        new Error(`${msg}file path ${pathDB} ` + 'does not exist'),
+      );
+    }
+  }
 }